+++ src/Main.java
@@ -1,139 +1,159 @@
-import java.util.ArrayList;
-import java.util.Arrays;
+import java.util.Vector;
 import java.util.concurrent.BrokenBarrierException;
 import java.util.concurrent.CyclicBarrier;
-import java.util.concurrent.Semaphore;
 
 class MainClass {
 
-    public Semaphore tunnelIngress = new Semaphore(2);
 
-        public static final int CARS_COUNT = 4;
-        public static void main(String[] args) {
 
-            CyclicBarrier prepToStart = new CyclicBarrier(CARS_COUNT);
 
-            System.out.println("ВАЖНОЕ ОБЪЯВЛЕНИЕ >>> Подготовка!!!");
-            Race race = new Race(new Road(60), new Tunnel(), new Road(40));
-            Car[] cars = new Car[CARS_COUNT];
-            for (int i = 0; i < cars.length; i++) {
-                cars[i] = new Car(race, 20 + (int) (Math.random() * 10));
-            }
-            for (int i = 0; i < cars.length; i++) {
-                new Thread(cars[i]).start();
+    public static final int CARS_COUNT = 4;
 
-            }
 
+    public static void main(String[] args) {
+        CyclicBarrier prepToStart = new CyclicBarrier(CARS_COUNT + 1);
+
+        System.out.println("ВАЖНОЕ ОБЪЯВЛЕНИЕ >>> Подготовка!!!");
+        Race race = new Race(new Road(60), new Tunnel(), new Road(40));
+        Car[] cars = new Car[CARS_COUNT];
+        for (int i = 0; i < cars.length; i++) {
+            cars[i] = new Car(race, 20 + (int) (Math.random() * 10), prepToStart);
+        }
+        for (int i = 0; i < cars.length; i++) {
+            new Thread(cars[i]).start();
+
+        }
+
+
+        try {
+            prepToStart.await();
+            prepToStart.await();
+            System.out.println("ВАЖНОЕ ОБЪЯВЛЕНИЕ >>> Гонка началась!!!");
+            prepToStart.await();
+
+        } catch (InterruptedException e) {
+            e.printStackTrace();
+        } catch (BrokenBarrierException e) {
+            e.printStackTrace();
+        }
+
+        System.out.println("ВАЖНОЕ ОБЪЯВЛЕНИЕ >>> Гонка закончилась!!!");
+        System.out.println("Победил " );//В процессе доработки
+    }
+
+}
+
+
+
+
+
+
+
+
+
+//    class Car implements Runnable {
+//
+//        private static int CARS_COUNT;
+//        static {
+//            CARS_COUNT = 0;
+//        }
+//        private Race race;
+//        private int speed;
+//        private String name;
+//        public String getName() {
+//            return name;
+//        }
+//        public int getSpeed() {
+//            return speed;
+//        }
+//        public Car(Race race, int speed) {
+//            this.race = race;
+//            this.speed = speed;
+//            CARS_COUNT++;
+//            this.name = "Участник #" + CARS_COUNT;
+//        }
+//        @Override
+//        public void run() {
 //            try {
-//                prepToStart.await();
-//            } catch (InterruptedException e) {
-//                e.printStackTrace();
-//            } catch (BrokenBarrierException e) {
+//                System.out.println(this.name + " готовится");
+//                Thread.sleep(500 + (int)(Math.random() * 800));
+//                System.out.println(this.name + " готов");
+//            } catch (Exception e) {
 //                e.printStackTrace();
 //            }
-            System.out.println("ВАЖНОЕ ОБЪЯВЛЕНИЕ >>> Гонка началась!!!");
-
-            System.out.println("ВАЖНОЕ ОБЪЯВЛЕНИЕ >>> Гонка закончилась!!!");
-
-
-
-
-        }
-
-
-
-
-    }
-    class Car implements Runnable {
-        private static int CARS_COUNT;
-        static {
-            CARS_COUNT = 0;
-        }
-        private Race race;
-        private int speed;
-        private String name;
-        public String getName() {
-            return name;
-        }
-        public int getSpeed() {
-            return speed;
-        }
-        public Car(Race race, int speed) {
-            this.race = race;
-            this.speed = speed;
-            CARS_COUNT++;
-            this.name = "Участник #" + CARS_COUNT;
-        }
-        @Override
-        public void run() {
-            try {
-                System.out.println(this.name + " готовится");
-                Thread.sleep(500 + (int)(Math.random() * 800));
-                System.out.println(this.name + " готов");
-            } catch (Exception e) {
-                e.printStackTrace();
-            }
-            for (int i = 0; i < race.getStages().size(); i++) {
-                race.getStages().get(i).go(this);
-            }
-        }
-    }
+//            for (int i = 0; i < race.getStages().size(); i++) {
+//                race.getStages().get(i).go(this);
+//            }
+//        }
+//    }
+//
+//
+//    abstract class Stage {
+//        protected int length;
+//        protected String description;
+//        public String getDescription() {
+//            return description;
+//        }
+//        public abstract void go(Car c);
+//    }
 
 
-    abstract class Stage {
-        protected int length;
-        protected String description;
-        public String getDescription() {
-            return description;
-        }
-        public abstract void go(Car c);
-    }
+//    class Road extends Stage {
+//        public Road(int length) {
+//            this.length = length;
+//            this.description = "Дорога " + length + " метров";
+//        }
+//        @Override
+//        public void go(Car c) {
+//            try {
+//                System.out.println(c.getName() + " начал этап: " + description);
+//                Thread.sleep(length / c.getSpeed() * 1000);
+//                System.out.println(c.getName() + " закончил этап: " + description);
+//            } catch (InterruptedException e) {
+//                e.printStackTrace();
+//            }
+//        }
+//    }
 
-    
-    class Road extends Stage {
-        public Road(int length) {
-            this.length = length;
-            this.description = "Дорога " + length + " метров";
-        }
-        @Override
-        public void go(Car c) {
-            try {
-                System.out.println(c.getName() + " начал этап: " + description);
-                Thread.sleep(length / c.getSpeed() * 1000);
-                System.out.println(c.getName() + " закончил этап: " + description);
-            } catch (InterruptedException e) {
-                e.printStackTrace();
-            }
-        }
-    }
-    class Tunnel extends Stage {
-        public Tunnel() {
-            this.length = 80;
-            this.description = "Тоннель " + length + " метров";
-        }
-        @Override
-        public void go(Car c) {
-            try {
-                try {
-                    System.out.println(c.getName() + " готовится к этапу(ждет): " + description);
 
-                    System.out.println(c.getName() + " начал этап: " + description);
-                    Thread.sleep(length / c.getSpeed() * 1000);
-                } catch (InterruptedException e) {
-                    e.printStackTrace();
-                } finally {
-                    System.out.println(c.getName() + " закончил этап: " + description);
-                }
-            } catch (Exception e) {
-                e.printStackTrace();
-            }
-        }
-    }
-    class Race {
-        private ArrayList<Stage> stages;
-        public ArrayList<Stage> getStages() { return stages; }
-        public Race(Stage... stages) {
-            this.stages = new ArrayList<>(Arrays.asList(stages));
-        }
-    }
+//    class Tunnel extends Stage {
+//
+//        public Semaphore tunnelIngress = new Semaphore(2);
+//
+//        public Tunnel() {
+//            this.length = 80;
+//            this.description = "Тоннель " + length + " метров";
+//        }
+//        @Override
+//        public void go(Car c) {
+//            try {
+//                try {
+//                    System.out.println(c.getName() + " готовится к этапу(ждет): " + description);
+//                    tunnelIngress.acquire();
+//                    System.out.println(c.getName() + " начал этап: " + description);
+//                    Thread.sleep(length / c.getSpeed() * 1000);
+//                } catch (InterruptedException e) {
+//                    e.printStackTrace();
+//                } finally {
+//                    System.out.println(c.getName() + " закончил этап: " + description);
+//                    tunnelIngress.release();
+//                }
+//            } catch (Exception e) {
+//                e.printStackTrace();
+//            }
+//        }
+//    }
+//
+
+//    class Race {
+//        private ArrayList<Stage> stages;
+//        public ArrayList<Stage> getStages() {
+//
+//            return stages;
+//
+//        }
+//        public Race(Stage... stages) {
+//            this.stages = new ArrayList<>(Arrays.asList(stages));
+//        }
+//    }
