+++ src/Race.java
@@ -1,0 +1,14 @@
+import java.util.ArrayList;
+import java.util.Arrays;
+
+class Race {
+    private ArrayList<Stage> stages;
+    public ArrayList<Stage> getStages() {
+
+        return stages;
+
+    }
+    public Race(Stage... stages) {
+        this.stages = new ArrayList<>(Arrays.asList(stages));
+    }
+}
