+++ src/Tunnel.java
@@ -1,0 +1,30 @@
+import java.util.concurrent.Semaphore;
+
+class Tunnel extends Stage {
+
+    public Semaphore tunnelIngress = new Semaphore(2);
+
+    public Tunnel() {
+        this.length = 80;
+        this.description = "Тоннель " + length + " метров";
+    }
+    @Override
+    public void go(Car c) {
+        try {
+            try {
+                System.out.println(c.getName() + " готовится к этапу(ждет): " + description);
+                tunnelIngress.acquire();
+                System.out.println(c.getName() + " начал этап: " + description);
+                Thread.sleep(length / c.getSpeed() * 1000);
+            } catch (InterruptedException e) {
+                e.printStackTrace();
+            } finally {
+                System.out.println(c.getName() + " закончил этап: " + description);
+                tunnelIngress.release();
+            }
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+    }
+}
+
